pyproject.toml
[tool.poetry]
name = "planes-around"
version = "0.1.0"
description = ""
authors = ["michaelfmnk <michael.fmnk@gmail.com>"]
readme = "README.md"
packages = [{include = "planes_around"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-telegram-bot = {extras = ["job-queue"], version = "^21.6"}
python-json-logger = "^2.0.7"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
